refactor: replace typeorm repo by postgres
