+++ packages/gateway/src/repositories/index.ts
@@ -1,3 +1,3 @@
-export * from './typeorm'
+export * from './postgres'
 export * from './l2'
 export * from './memory'
